feat: add gc mock, prettify gc vis, add legends, enter on works with hw keyboard, fix vfs_write data aggregation and make it responsive to interval, ktfmt

+++ frontend/app/src/main/java/de/amosproj3/ziofa/ui/visualization/data/GraphedData.kt
@@ -20,7 +20,7 @@ sealed class GraphedData {
     ) : GraphedData()
 
     data class HistogramData(
-        val seriesData: ImmutableList<Pair<ULong, ULong>>,
+        val seriesData: ImmutableList<Pair<ULong, Double>>,
         val metaData: ChartMetadata,
     ) : GraphedData()
